IOTS-296: Generalizing policy UI for policy create wizard
This commit is done to seperate policy related client side logic and UIs from device type specific logics.
This is a part of creating self contained plugins  for device types.

This commit generalizes the policy creation wizard in the CDMF core and removes device type specific UIs and client side logic from the core. Also, this introduces a generic policy configuration UI in a new unit.