Fixed JWT properties reading issue
Previously JWT properties file was read to get the defult-jwt-client properties. If someone wants to avoid reading the default keystores, it can only be done by uploading a jwt properties to registry and reading from itself.
This fix changes the default properties reading mechanism to get the properties from the file and creates a jwt client without being the default