Checking if the user store manager class is a sub class of the default
This commit is to modify the current logic which checks if the UserStoreManager class is equal to the JDBCUserStoreManager class to support subclasses of the JDBCUserStoreManager class to return correct total user count.

In this commit, instead of comparing the string name, tests the UserStoreManager instance is a type of JDBCUserStoreManager.

Resolves wso2/product-iots#1556