Fixing policy view interface not loading issue
When a policy is created and viewed the policy using the policy view icon the page keeps loading sometimes due to undefined deviceGroups variable inside the policyPayloadObj. 

This commit fixes this issue by improving the code where the null check has been made to check if the  deviceGroup is null. Also made few improvements in the code where the checks have been made.

Resolves wso2/product-iots#1681