Adding entity application mapping to the database
When an application installation request is made other than adding an application install operation a mapping between the devices and the installed applications are kept in the database. In a case where an enterprise wide installation happens a mapping is made with the respective entity and the application.

This commit implement the mapping function.

Related to issue wso2/product-iots#1688