Add DEP management icon to navmenu
The DEP management page is formed once the iOS plugin has been deployed
within the IoT Home. Till then, any functionality with regards to DEP
management is not present within the server.
An icon in the navmenu allows users to access the DEP management page.

This commit includes the feature where the navmenu detects if the iOS
plugin has been deployed and accordingly displays the icon to access
the DEP management page.