Make proxy configuration optional
As per the current configurations it is mandatory to set a value for http.nonProxyHosts property in order to server to run on 'change-ip' mode. With this modification it is made optional where even without this property IoT server can be run on 'change-ip' mode. 'Change-ip' mode is where server is configured to run on a different IP other than localhost